pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mss-aggregate"
version = "0.1.0"
description = "Aggregate multiple MSS datasets into unified stem folders"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "soundfile>=0.12",
    "numpy>=1.21",
    "pyyaml>=6.0",
    "unidecode>=1.3",
    "tqdm>=4.60",
    "click>=8.0",
]

[project.optional-dependencies]
moisesdb = ["moisesdb>=0.2"]
loudness = ["pyloudnorm>=0.1"]
rich = ["rich>=12.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
mss-aggregate = "mss_aggregate.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
